[#85019] Constrain C# language version to 7.2
It is to prevent compilation error on C# 14 (.NET 10) due to compiler-generated backing field being exposed with "field" keywoard. Therefore "field" becomes reserved in the context of property setter/getter. To avoid modifications to the existing code, we explicitly tell the compiler to not use newer language features. That constraint can be lifted in the future after small modifications to the code. Reference: https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/field